pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[project]
name = 'midiscripter'
version = '0.8'
description = 'Framework for scripting MIDI, Open Sound Control, keyboard and mouse input and output'
readme = 'README.md'
requires-python = '>=3.11'
license = 'LGPL-3.0-or-later'
license-files = ['LICENSE']
keywords = ['MIDI', 'OSC', 'script', 'automation', 'input']
authors = [{name = 'Maboroshy'}]
classifiers = [
  'Development Status :: 4 - Beta',
  'Programming Language :: Python :: 3.11',
  'Programming Language :: Python :: 3.12',
  'Programming Language :: Python :: 3.13',
  'License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)',
  'Topic :: Multimedia :: Sound/Audio',
  'Topic :: Multimedia :: Sound/Audio :: MIDI',
  'Topic :: Software Development :: Libraries :: Python Modules',
]
dependencies = [
  'colorama ~=0.4',  # BSD
  'python-rtmidi ~=1.5',  # MIT
  'python-osc ~=1.7',  # Unlicense
  'pynput ~=1.7',  # LGPL
  'watchdog ~=4.0',  # Apache
  'PySide6_Essentials >6.4',  # LGPL
  'pywin32; platform_system == "Windows"',  # BSD
]

[project.optional-dependencies]
dev = ['ruff']
doc = ['mkdocs-material', 'mkdocstrings', 'griffe-inherited-docstrings', 'black']

[tool.hatch.build.targets.sdist]
only-include = ['midiscripter']

[project.gui-scripts]
midiscripter = 'midiscripter:start_gui'

[project.urls]
Documentation = 'https://maboroshy.github.io/midi-scripter'
Issues = 'https://github.com/Maboroshy/midi-scripter/issues'
Source = 'https://github.com/Maboroshy/midi-scripter'

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ['B', 'C', 'E', 'F', 'PERF', 'TCH', 'SIM', 'UP', 'ANN001', 'ANN201', 'ANN202']
ignore = ['E501', 'F401', 'F405', 'F403', 'SIM105']

[tool.ruff.format]
quote-style = 'single'
